Se agrego tests al 100% de RestriccionDeDiasDeSemanaTest
